Apply fix from load_sp to load_sp_dfs

+++ src/implementations/algorithms/load_sp_dfs.h
@@ -22,8 +22,8 @@
 typedef struct {
     uint8_t  rem_bits;   // remaining bits
     uint8_t  first_diff; // first difference
-    uint32_t in_idx;     // input_index
-    uint32_t out_idx;    // output_index
+    uint64_t in_idx;     // input_index
+    uint64_t out_idx;    // output_index
 } MergeOp;
 
 /*
@@ -35,7 +35,7 @@ static MergeOp *load_schedule(int num_bits, size_t *out_count) {
     // 1) compute total
     printf("loaded");
     size_t total = (1 << num_bits)-1;
-    
+
 
     MergeOp *ops = (MergeOp *)malloc(sizeof *ops * total);
     if (!ops) {
@@ -56,8 +56,8 @@ static MergeOp *load_schedule(int num_bits, size_t *out_count) {
     size_t idx = 0;
     while (idx < total) {
         int rem, diff;
-        uint32_t in, out;
-        int r = fscanf(f, "%d %d %u %u\n", &rem, &diff, &in, &out);
+        uint64_t in, out;
+        int r = fscanf(f, "%d %d %llu %llu\n", &rem, &diff, &in, &out);
         if (r == EOF) break;
         if (r != 4) {
             // skip malformed/blank/comment lines
